+++ pysui/sui/sui_clients/sync_client.py
@@ -17,7 +17,7 @@
 from typing import Any, Union
 from json import JSONDecodeError
 import httpx
-from pysui.sui.sui_clients.common import _ClientMixin, SuiRpcResult
+from pysui.sui.sui_clients.common import _ClientMixin, PreExecutionResult, SuiRpcResult
 from pysui.sui.sui_types.scalars import ObjectID, SuiInteger, SuiTxBytes, SuiString
 from pysui.sui.sui_types.address import SuiAddress
 from pysui.sui.sui_types.collections import SuiArray, SuiMap
@@ -106,7 +106,10 @@ class SuiClient(_ClientMixin):
                 result = result.result_data
                 if "error" in result:
                     return SuiRpcResult(False, result["error"]["message"], None)
-                result = SuiRpcResult(True, None, (builder.authority, SuiTxBytes(result["result"]["txBytes"])))
+                result = SuiRpcResult(
+                    True, None, PreExecutionResult(builder.authority, builder.handle_return(result["result"]))
+                )
+                #    SuiTxBytes(result["result"]["txBytes"])))
             return result
         return SuiRpcResult(False, "execute_no_sign is used only with transaction types")
 
@@ -145,8 +148,7 @@ class SuiClient(_ClientMixin):
         if builder.txn_required:
             result = self.execute_no_sign(builder)
             if result.is_ok():
-                _, tx_bytes = result.result_data
-                result = self.execute(DryRunTransaction(tx_bytes=tx_bytes))
+                result = self.execute(DryRunTransaction(tx_bytes=result.result_data.tx_bytes))
             return result
         return SuiRpcResult(False, "dry_run is used only with transaction types")
 
@@ -156,10 +158,9 @@ class SuiClient(_ClientMixin):
         """."""
         result = self.execute_no_sign(builder)
         if result.is_ok():
-            tx_bytes = result.result_data[1]
-            exec_builder = self.sign_for_execution(tx_bytes, builder, additional_signers)
+            exec_builder = self.sign_for_execution(result.result_data.tx_bytes, builder, additional_signers)
             result = self._execute(exec_builder)
-            if result.is_ok() and "error" not in result.result_data["result"]:
+            if result.is_ok() and "error" not in result.result_data:
                 # print(result.result_data["result"])
                 result = SuiRpcResult(True, None, exec_builder.handle_return(result.result_data["result"]))
         return result
@@ -224,12 +225,15 @@ class SuiClient(_ClientMixin):
         """
         for_address = for_address or self.config.active_address
         try:
+            print(self.config.faucet_url)
+            print({"FixedAmountRequest": {"recipient": f"{for_address}"}})
             result = self._client.post(
                 self.config.faucet_url,
                 headers=GetObjectsOwnedByAddress(for_address).header,
                 json={"FixedAmountRequest": {"recipient": f"{for_address}"}},
             ).json()
             if result["error"] is None:
+                print(result)
                 return SuiRpcResult(True, None, FaucetGasRequest.from_dict(result))
             return SuiRpcResult(False, result["error"])
         except JSONDecodeError as jexc: